Update Clear System

+++ src/main/java/bid/ApixTeam/bot/utils/api/MessageManager.java
@@ -61,7 +61,7 @@ public class MessageManager {
         int counter = 1;
         for(Message iMessage : messages) {
             if(counter > amount) break;
-            if(iMessage.getAuthor().getId().equals(user.getId())) {
+            if(iMessage.getAuthor().getId().equals(user.getId()) && !(iMessage.getContent().startsWith("!"))) {
                 Consumer<Void> callback = (response) -> System.out.printf("[%s] [INFO] [API] [%s:%s] Bulk message delete. %n", new Date(), user.getName(), textChannel.getName());
                 textChannel.deleteMessageById(iMessage.getId()).queue(callback);
                 counter++;
